fix: Tricky Norm fix (#38)
* merge two branch with completely different history to avoid huge rebase

* fix: Make after tricky Norm fix

---------
