+++ command_args.c
@@ -1,0 +1,80 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   command_args.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mwiecek <mwiecek@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/23 17:58:27 by zslowian          #+#    #+#             */
+/*   Updated: 2025/04/28 19:57:05 by mwiecek          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	create_args_default_mode(t_token **token_node, t_command *last_cmd)
+{
+	int		i;
+	t_token	*temp;
+
+	i = 0;
+	temp = *token_node;
+	last_cmd->args_size = count_arguments(temp) + 2;
+	last_cmd->args = malloc(sizeof(char *) * (last_cmd->args_size));
+	if (!last_cmd->args)
+		return (1);
+	temp = *token_node;
+	i = 0;
+	last_cmd->args[i] = ft_strdup(last_cmd->command);
+	i++;
+	while (temp->type == WORD || temp->type == VAR)
+	{
+		last_cmd->args[i] = ft_strdup(temp->str);
+		i++;
+		temp = temp->next;
+	}
+	last_cmd->args[i] = NULL;
+	*token_node = temp;
+	return (0);
+}
+
+char	*join_vars(t_token **token_node)
+{
+	t_token	*temp;
+	char	*str;
+
+	temp = *token_node;
+	str = ft_strdup(temp->str);
+	while (temp->type == VAR && temp->next->type == VAR
+		&& temp->next->join == true)
+	{
+		str = ft_strjoin(str, temp->next->str);
+		temp = temp->next;
+	}
+	*token_node = temp;
+	return (str);
+}
+
+char	**copy_in_new_tab(int len, char **new_tab,
+			t_command *last_cmd, t_token *tmp)
+{
+	int	i;
+
+	i = 0;
+	while (i < len)
+	{
+		new_tab[i] = last_cmd->args[i];
+		i++;
+	}
+	while (tmp->type == WORD || tmp->type == VAR)
+	{
+		if (tmp->join == true)
+			new_tab[i] = join_vars(&tmp);
+		else
+			new_tab[i] = ft_strdup(tmp->str);
+		i++;
+		tmp = tmp->next;
+	}
+	new_tab[i] = NULL;
+	return (new_tab);
+}
