+++ command_vars.c
@@ -1,0 +1,143 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   command_vars.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mwiecek <mwiecek@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/23 17:58:27 by zslowian          #+#    #+#             */
+/*   Updated: 2025/04/28 18:43:36 by mwiecek          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+static char	**copy_default_in_new_tab(
+	int len, char **new_tab, t_command *last_cmd, t_token **tk_node)
+{
+	int		i;
+	t_token	*temp;
+
+	i = 0;
+	temp = *tk_node;
+	while (i < len)
+	{
+		new_tab[i] = last_cmd->args[i];
+		i++;
+	}
+	while (temp->type == WORD || temp->type == VAR)
+	{
+		new_tab[i] = ft_strdup(temp->str);
+		i++;
+		temp = temp->next;
+	}
+	new_tab[i] = NULL;
+	return (new_tab);
+}
+
+int	add_args_default_mode(t_token **token_node, t_command *last_cmd)
+{
+	int		i;
+	int		len;
+	char	**new_tab;
+	t_token	*temp;
+
+	i = 0;
+	temp = *token_node;
+	while (temp->type == WORD || temp->type == VAR)
+	{
+		i++;
+		temp = temp->next;
+	}
+	len = 0;
+	while (last_cmd->args[len])
+		len++;
+	new_tab = malloc(sizeof(char *) * (i + len + 1));
+	if (!new_tab)
+		return (1);
+	new_tab = copy_default_in_new_tab(len, new_tab, last_cmd, token_node);
+	free(last_cmd->args);
+	last_cmd->args = new_tab;
+	*token_node = temp;
+	return (0);
+}
+
+/*
+**  This function fills the arguments in the command structure (command->args)
+**  It has two modes:
+**    - The "echo mode" if the command is the builtin "echo"
+**    - The default mode for all the other cases
+*/
+
+int	fill_args(t_token **token_node, t_command *last_cmd)
+{
+	if (!ft_strcmp(last_cmd->command, ECHO))
+	{
+		if (!(last_cmd->args))
+			return (create_args_echo_mode(token_node, last_cmd));
+		else
+			return (add_args_echo_mode(token_node, last_cmd));
+	}
+	else
+	{
+		if (last_cmd && !(last_cmd->args))
+			return (create_args_default_mode(token_node, last_cmd));
+		else
+			return (add_args_default_mode(token_node, last_cmd));
+	}
+	return (0);
+}
+
+static void	split_var_cmd_token(t_command *last_cmd, char *cmd_str)
+{
+	t_token		*new_tokens;
+	t_token		*tmp;
+	char		**strs;
+	int			i;
+
+	new_tokens = NULL;
+	strs = ft_split(cmd_str, ' ');
+	if (!strs)
+		return ;
+	last_cmd->command = ft_strdup(strs[0]);
+	if (strs[1])
+		new_tokens = lst_new_token(ft_strdup(strs[1]), NULL, WORD, DEFAULT);
+	tmp = new_tokens;
+	i = 1;
+	while (strs[++i])
+		lst_add_back_token(&new_tokens,
+			lst_new_token(ft_strdup(strs[i]), NULL, WORD, DEFAULT));
+	lst_add_back_token(&new_tokens,
+		lst_new_token(NULL, NULL, END, DEFAULT));
+	fill_args(&new_tokens, last_cmd);
+	lstclear_token(&tmp, &free_ptr);
+	free_str_tab(strs);
+}
+
+void	parse_word(t_command **cmd, t_token **token_lst, t_global *g)
+{
+	t_token		*temp;
+	t_command	*last_cmd;
+
+	temp = *token_lst;
+	while (temp->type == WORD || temp->type == VAR)
+	{
+		last_cmd = lst_last_cmd(*cmd);
+		if (temp->prev == NULL || (temp->prev && temp->prev->type == PIPE)
+			|| last_cmd->command == NULL)
+		{
+			if (temp->type == VAR && contains_space(temp->str))
+				split_var_cmd_token(last_cmd, temp->str);
+			else
+			{
+				last_cmd->command = ft_strdup(temp->str);
+				last_cmd->is_builtin = ft_is_our_builtin(last_cmd->command, g);
+				ft_is_status_request(temp, last_cmd);
+			}
+			temp = temp->next;
+		}
+		else
+			fill_args(&temp, last_cmd);
+	}
+	*token_lst = temp;
+}
