+++ commands_append_pipe.c
@@ -1,0 +1,76 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   commands_append_pipe.c                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: zslowian <zslowian@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/23 17:58:27 by zslowian          #+#    #+#             */
+/*   Updated: 2025/04/23 18:20:04 by zslowian         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+/*
+	***APPEND***
+	Redirection of output in append mode causes the file whose name results
+	from the expansion of word to be opened for appending on file descriptor n,
+	or the standard output (fd 1) if n is not specified.
+	If the file does not exist it is created.
+
+	The general format for appending output is: [n]>>word.
+
+*/
+
+/* open_outfile_append:
+*	Opens an outfile in append mode. If an outfile was already set, frees it
+*	and overwrites it. If a previous infile or outfile open failed (file does
+*	not exist or permission denied), does not open any further output file.
+*
+*	Ex.:
+*		echo hello > forbidden_file >> test
+*		echo hello >> forbidden_file >> test
+*		< forbidden_file cat >> test
+*	In these 3 cases, the test file should not be opened or created.
+*/
+static void	open_outfile_append(t_io_fds *io, char *file, char *var_filename)
+{
+	if (!remove_old_file_ref(io, false))
+		return ;
+	io->outfile = ft_strdup(file);
+	if (io->outfile && io->outfile[0] == '\0' && var_filename)
+	{
+		errmsg_cmd(var_filename, NULL, "ambiguous redirect", false);
+		return ;
+	}
+	io->fd_out = open(io->outfile, O_WRONLY | O_CREAT | O_APPEND, 0664);
+	if (io->fd_out == -1)
+		errmsg_cmd(io->outfile, NULL, strerror(errno), false);
+}
+
+void	parse_append(t_command **last_cmd, t_token **token_lst)
+{
+	t_token		*temp;
+	t_command	*cmd;
+
+	temp = *token_lst;
+	cmd = lst_last_cmd(*last_cmd);
+	init_io(cmd);
+	open_outfile_append(cmd->io_fds, temp->next->str, temp->next->str_backup);
+	if (temp->next->next)
+		temp = temp->next->next;
+	else
+		temp = temp->next;
+	*token_lst = temp;
+}
+
+void	parse_pipe(t_command **cmd, t_token **token_lst)
+{
+	t_command	*last_cmd;
+
+	last_cmd = lst_last_cmd(*cmd);
+	last_cmd->pipe_output = true;
+	lst_add_back_cmd(&last_cmd, lst_new_cmd(false));
+	*token_lst = (*token_lst)->next;
+}
