+++ commands_heredoc.c
@@ -1,0 +1,124 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   commands_heredoc.c                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mwiecek <mwiecek@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/23 17:58:27 by zslowian          #+#    #+#             */
+/*   Updated: 2025/04/28 22:01:33 by mwiecek          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+static char	*make_str_from_tab(char **tab)
+{
+	char	*str;
+	char	*tmp;
+	int		i;
+
+	i = -1;
+	tmp = "";
+	while (tab[++i])
+	{
+		if (i == 0)
+			str = ft_strdup(tab[0]);
+		else
+		{
+			str = ft_strjoin(tmp, tab[i]);
+			free_ptr((void **)&tmp);
+		}
+		if (tab[i + 1])
+		{
+			tmp = str;
+			str = ft_strjoin(tmp, " ");
+			free_ptr((void **)&tmp);
+		}
+	}
+	free_str_tab(tab);
+	return (str);
+}
+
+static char	*get_expanded_var_line(t_global *global, char *line)
+{
+	char	**words;
+	int		i;
+
+	words = ft_split(line, ' ');
+	if (!words)
+		return (NULL);
+	i = 0;
+	while (words[i])
+	{
+		if (ft_strchr(words[i], '$'))
+		{
+			words[i] = var_expander_heredoc(global, words[i]);
+			if (!words[i])
+				return (NULL);
+		}
+		i++;
+	}
+	return (make_str_from_tab(words));
+}
+
+static bool	evaluate_heredoc_line(t_global *global, char **line,
+									t_io_fds *io, bool *ret)
+{
+	if (*line == NULL)
+	{
+		errmsg_cmd("warning", "here-document delimited by end-of-file: wanted",
+			io->heredoc_delimiter, true);
+		*ret = true;
+		return (false);
+	}
+	if (ft_strcmp(*line, io->heredoc_delimiter) == 0)
+	{
+		*ret = true;
+		return (false);
+	}
+	if (io->heredoc_quotes == false && ft_strchr(*line, '$'))
+	{
+		*line = get_expanded_var_line(global, *line);
+		if (!(*line))
+		{
+			free_ptr((void **)&*line);
+			*ret = false;
+			return (false);
+		}
+	}
+	return (true);
+}
+
+bool	fill_heredoc(t_global *global, t_io_fds *io, int fd)
+{
+	char	*line;
+	bool	ret;
+
+	ret = false;
+	line = NULL;
+	while (1)
+	{
+		set_signals_interactive();
+		line = readline(">");
+		set_signals_noninteractive();
+		if (!evaluate_heredoc_line(global, &line, io, &ret))
+			break ;
+		ft_putendl_fd(line, fd);
+		free_ptr((void **)&line);
+	}
+	free_ptr((void **)&line);
+	return (ret);
+}
+
+bool	get_heredoc(t_global *global, t_io_fds *io)
+{
+	int		tmp_fd;
+	bool	ret;
+
+	ret = true;
+	tmp_fd = open(io->infile, O_CREAT | O_WRONLY | O_TRUNC, 0644);
+	ret = fill_heredoc(global, io, tmp_fd);
+	close(tmp_fd);
+	return (ret);
+}
