+++ parser/characters.c
@@ -53,3 +53,46 @@ int	which_separator(char *str, int i)
 	else
 		return (0);
 }
+
+bool	quotes_in_string(char *str)
+{
+	int	i;
+
+	i = 0;
+	while (str[i])
+	{
+		if (str[i] == '\'' || str[i] == '\"')
+			return (true);
+		i++;
+	}
+	return (false);
+}
+
+int	count_len(char *str, int count, int i)
+{
+	int	status;
+
+	status = 0;
+	while (str[i])
+	{
+		if ((str[i] == '\'' || str[i] == '\"') && status == DEFAULT)
+		{
+			if (str[i] == '\'')
+				status = SQUOTE;
+			if (str[i] == '\"')
+				status = DQUOTE;
+			i++;
+			continue ;
+		}
+		else if ((str[i] == '\'' && status == SQUOTE)
+			|| (str[i] == '\"' && status == DQUOTE))
+		{
+			status = DEFAULT;
+			i++;
+			continue ;
+		}
+		count++;
+		i++;
+	}
+	return (count + 1);
+}
