+++ parser/nodes.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   nodes.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zslowian <zslowian@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mwiecek <mwiecek@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 18:12:04 by zslowian          #+#    #+#             */
-/*   Updated: 2025/04/23 18:12:49 by zslowian         ###   ########.fr       */
+/*   Updated: 2025/04/28 18:36:34 by mwiecek          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
