+++ parser/signals.c
@@ -47,7 +47,6 @@ void	set_signals_interactive(void)
 void	signal_print_newline(int signal)
 {
 	(void)signal;
-	write(1, "\n", 1);
 	rl_on_new_line();
 }
