+++ tokens_list.c
@@ -1,0 +1,82 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   tokens_list.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mwiecek <mwiecek@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/23 17:58:27 by zslowian          #+#    #+#             */
+/*   Updated: 2025/04/28 21:35:59 by mwiecek          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+t_token	*lst_new_token(char *str, char *str_backup, int type, int status)
+{
+	t_token	*new_node;
+
+	new_node = malloc(sizeof(t_token) * 1);
+	if (!(new_node))
+		return (NULL);
+	new_node->str = str;
+	new_node->str_backup = str_backup;
+	new_node->var_exists = false;
+	new_node->type = type;
+	new_node->status = status;
+	new_node->join = false;
+	new_node->prev = NULL;
+	new_node->next = NULL;
+	return (new_node);
+}
+
+void	lst_add_back_token(t_token **alst, t_token *new_node)
+{
+	t_token	*start;
+
+	start = *alst;
+	if (start == NULL)
+	{
+		*alst = new_node;
+		return ;
+	}
+	if (alst && *alst && new_node)
+	{
+		while (start->next != NULL)
+			start = start->next;
+		start->next = new_node;
+		new_node->prev = start;
+	}
+}
+
+void	lstdelone_token(t_token *lst, void (*del)(void **))
+{
+	if (del && lst && lst->str)
+	{
+		(*del)((void **)&lst->str);
+		lst->str = NULL;
+	}
+	if (del && lst && lst->str_backup)
+	{
+		(*del)((void **)&lst->str_backup);
+		lst->str_backup = NULL;
+	}
+	if (lst->prev)
+		lst->prev->next = lst->next;
+	if (lst->next)
+		lst->next->prev = lst->prev;
+	free_ptr((void **)&lst);
+}
+
+void	lstclear_token(t_token **lst, void (*del)(void **))
+{
+	t_token	*tmp;
+
+	tmp = NULL;
+	while (*lst != NULL)
+	{
+		tmp = (*lst)->next;
+		lstdelone_token(*lst, del);
+		*lst = tmp;
+	}
+}
